// frontend/src/newRedux/types/userTypes.ts
export interface UserData {
    id: string;
    username: string;
    firstName: string;
    lastName: string;
    email: string;
    avatar?: string;
    bio?: string;
    isAdmin?: boolean;
    videos?: string[];
    photos?: string[];
    favMountains?: string[];
};

export interface NewUserData {
    username: string;
    password: string;
    firstName: string;
    lastName: string;
    email: string;
};

export interface LoginData {
    username: string;
    password: string;
}

export interface UpdateUserData {
    username?: string;
    firstName?: string;
    lastName?: string;
    password?: string;
    email?: string;
    avatar?: string;
    bio?: string;
    videos?: string[];
    photos?: string[];
    favMountains?: string[];
};

export interface MakeUserAdmin {
    username: string;
    isAdmin: boolean;
}

export const FETCH_USER_DATA_REQUEST = 'FETCH_USER_DATA_REQUEST';
export const FETCH_USER_DATA_SUCCESS = 'FETCH_USER_DATA_SUCCESS';
export const FETCH_USER_DATA_FAILURE = 'FETCH_USER_DATA_FAILURE';
export const SEND_NEW_USER_DATA_REQUEST = 'SEND_NEW_USER_DATA_REQUEST';
export const SEND_NEW_USER_DATA_SUCCESS = 'SEND_NEW_USER_DATA_SUCCESS';
export const SEND_NEW_USER_DATA_FAILURE = 'SEND_NEW_USER_DATA_FAILURE';
export const UPDATE_USER_DATA_REQUEST = 'UPDATE_USER_DATA_REQUEST';
export const UPDATE_USER_DATA_SUCCESS = 'UPDATE_USER_DATA_SUCCESS';
export const UPDATE_USER_DATA_FAILURE = 'UPDATE_USER_DATA_FAILURE';
export const DELETE_USER_REQUEST = 'DELETE_USER_REQUEST';
export const DELETE_USER_SUCCESS = 'DELETE_USER_SUCCESS';
export const DELETE_USER_FAILURE = 'DELETE_USER_FAILURE';
export const SEND_LOGIN_DATA_REQUEST = 'SEND_LOGIN_DATA_REQUEST';
export const SEND_LOGIN_DATA_SUCCESS = 'SEND_LOGIN_DATA_SUCCESS';
export const SEND_LOGIN_DATA_FAILURE = 'SEND_LOGIN_DATA_FAILURE';
export const SET_TOKEN = 'SET_TOKEN';
export const LOGOUT_USER = 'LOGOUT_USER';
export const MAKE_ADMIN_DATA_REQUEST = 'MAKE_ADMIN_DATA_REQUEST';
export const MAKE_ADMIN_DATA_SUCCESS = 'MAKE_ADMIN_DATA_SUCCESS';
export const MAKE_ADMIN_DATA_FAILURE = 'MAKE_ADMIN_DATA_FAILURE';